==> Part1_Data Cleaning.ipynb <==
# Importing necessary modules
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import datetime as dt
import warnings
warnings.filterwarnings("ignore")

# Reading Card Transactions Raw Data
data = pd.read_excel('card transactions.xlsx')

## Exclusions

#Retaining only P transaction type
data = data[data['Transtype'] == 'P']
data.head()

#Removing the largest transaction (It is in Pesos)
data = data[data['Amount'] < 3000000]
data.info()

#% of missing records
data.count() * 100 /len(data)

#checking nulls
data.isnull().sum()

## Data Cleaning

# Replacing 0s with nan
data['Merchnum'] = data['Merchnum'].replace(np.nan,'0')
data['Merch state'] = data['Merch state'].replace(np.nan,'0')
data['Merch zip'] = data['Merch zip'].replace(np.nan,0)
data.isnull().sum()

### Filling Merchnum

data['Merchnum'].replace({'0':np.nan},inplace = True)
data.isnull().sum()

data['Merchnum']=data.groupby(['Merch description','Merch zip'])['Merchnum'].transform(lambda x:x.fillna(x.mode()[0]) if not x.mode().empty else 'Empty')
data['Merchnum'].replace({'Empty':np.nan},inplace = True)
data['Merchnum']=data.groupby(['Merch description','Merch state'])['Merchnum'].transform(lambda x:x.fillna(x.mode()[0]) if not x.mode().empty else 'Empty')
data['Merchnum'].replace({'Empty':np.nan},inplace = True)
data['Merchnum']=data.groupby('Merch description')['Merchnum'].transform(lambda x:x.fillna(x.mode()[0]) if not x.mode().empty else 'Empty')
data['Merchnum']=np.where(data['Merchnum']=='Empty',data['Recnum']*(-1),data['Merchnum'])
data.isnull().sum()

### Filling State & Zip

data['Merch state'].replace({'0':np.nan},inplace = True)
data['Merch zip'].replace({0:np.nan},inplace = True)
data.isnull().sum()

data['Merch state']=data.groupby('Merchnum')['Merch state'].transform(lambda x:x.fillna(x.mode()[0]) if not x.mode().empty else 'Empty')
data['Merch zip']=data.groupby('Merchnum')['Merch zip'].transform(lambda x:x.fillna(x.mode()[0]) if not x.mode().empty else 'Empty')
data['Merch state']=np.where(data['Merch state']=='Empty',data['Recnum']*(-1),data['Merch state'])
data['Merch zip']=np.where(data['Merch zip']=='Empty',data['Recnum']*(-1),data['Merch zip'])
data.isnull().sum()

data.shape

#data.to_csv("Cleaned Data.csv")
