==> Part2_Variable Creation.ipynb <==
# Importing necessary modules
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import datetime as dt
import warnings
warnings.filterwarnings("ignore")

# Reading Card Transactions Raw Data
df1 = pd.read_csv('Cleaned Data.csv')

df1.shape

# Converting to right format
df1['Date'] = pd.to_datetime(df1['Date'])
df1.info()

df1['Cardnum'] = df1['Cardnum'].astype(str)
df1['Merch zip'] = df1['Merch zip'].astype(str)
df1.info()

## Variable Creation

df1 = df1.set_index('Date')
df1.head()

#Calculates all the amount variables (200)

df1 = df1.reset_index()
df1 = df1.sort_values(by = ['Cardnum', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
      df1['mean_amt_{}_{}'.format('card',days)] = getattr(df1_index.groupby('Cardnum')['Amount'].rolling(days),'mean')().values
      df1['max_amt_{}_{}'.format('card',days)] = getattr(df1_index.groupby('Cardnum')['Amount'].rolling(days),'max')().values
      df1['median_amt_{}_{}'.format('card',days)] = getattr(df1_index.groupby('Cardnum')['Amount'].rolling(days),'median')().values
      df1['sum_amt_{}_{}'.format('card',days)] = getattr(df1_index.groupby('Cardnum')['Amount'].rolling(days),'sum')().values
df1.sort_values(by = ['Recnum'],inplace = True)


df1 = df1.reset_index(drop = True)
df1 = df1.sort_values(by = ['Merchnum', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
      df1['mean_amt_{}_{}'.format('merch',days)] = getattr(df1_index.groupby('Merchnum')['Amount'].rolling(days),'mean')().values
      df1['max_amt_{}_{}'.format('merch',days)] = getattr(df1_index.groupby('Merchnum')['Amount'].rolling(days),'max')().values
      df1['median_amt_{}_{}'.format('merch',days)] = getattr(df1_index.groupby('Merchnum')['Amount'].rolling(days),'median')().values
      df1['sum_amt_{}_{}'.format('merch',days)] = getattr(df1_index.groupby('Merchnum')['Amount'].rolling(days),'sum')().values
df1.sort_values(by = ['Recnum'],inplace = True)


df1 = df1.reset_index(drop = True)
df1 = df1.sort_values(by = ['Cardnum','Merchnum', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
      df1['mean_amt_{}_{}'.format('cardmerch',days)] = getattr(df1_index.groupby(['Cardnum','Merchnum'])['Amount'].rolling(days),'mean')().values
      df1['max_amt_{}_{}'.format('cardmerch',days)] = getattr(df1_index.groupby(['Cardnum','Merchnum'])['Amount'].rolling(days),'max')().values
      df1['median_amt_{}_{}'.format('cardmerch',days)] = getattr(df1_index.groupby(['Cardnum','Merchnum'])['Amount'].rolling(days),'median')().values
      df1['sum_amt_{}_{}'.format('cardmerch',days)] = getattr(df1_index.groupby(['Cardnum','Merchnum'])['Amount'].rolling(days),'sum')().values
df1.sort_values(by = ['Recnum'],inplace = True)


df1 = df1.reset_index(drop = True)
df1 = df1.sort_values(by = ['Cardnum','Merch zip', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
      df1['mean_amt_{}_{}'.format('cardzip',days)] = getattr(df1_index.groupby(['Cardnum','Merch zip'])['Amount'].rolling(days),'mean')().values
      df1['max_amt_{}_{}'.format('cardzip',days)] = getattr(df1_index.groupby(['Cardnum','Merch zip'])['Amount'].rolling(days),'max')().values
      df1['median_amt_{}_{}'.format('cardzip',days)] = getattr(df1_index.groupby(['Cardnum','Merch zip'])['Amount'].rolling(days),'median')().values
      df1['sum_amt_{}_{}'.format('cardzip',days)] = getattr(df1_index.groupby(['Cardnum','Merch zip'])['Amount'].rolling(days),'sum')().values
df1.sort_values(by = ['Recnum'],inplace = True)


df1 = df1.reset_index(drop = True)
df1 = df1.sort_values(by = ['Cardnum','Merch state', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
      df1['mean_amt_{}_{}'.format('cardstate',days)] = getattr(df1_index.groupby(['Cardnum','Merch state'])['Amount'].rolling(days),'mean')().values
      df1['max_amt_{}_{}'.format('cardstate',days)] = getattr(df1_index.groupby(['Cardnum','Merch state'])['Amount'].rolling(days),'max')().values
      df1['median_amt_{}_{}'.format('cardstate',days)] = getattr(df1_index.groupby(['Cardnum','Merch state'])['Amount'].rolling(days),'median')().values
      df1['sum_amt_{}_{}'.format('cardstate',days)] = getattr(df1_index.groupby(['Cardnum','Merch state'])['Amount'].rolling(days),'sum')().values
df1.sort_values(by = ['Recnum'],inplace = True)
        
for days in ['1d', '3d', '7d','14d','30d']:
      df1['ratio_actual_mean_amt_{}_{}'.format('card',days)] = df1['Amount']/df1['mean_amt_{}_{}'.format('card',days)]
      df1['ratio_actual_mean_amt_{}_{}'.format('merch',days)] = df1['Amount']/df1['mean_amt_{}_{}'.format('merch',days)]
      df1['ratio_actual_mean_amt_{}_{}'.format('cardmerch',days)] = df1['Amount']/df1['mean_amt_{}_{}'.format('cardmerch',days)]
      df1['ratio_actual_mean_amt_{}_{}'.format('cardzip',days)] = df1['Amount']/df1['mean_amt_{}_{}'.format('cardzip',days)]
      df1['ratio_actual_mean_amt_{}_{}'.format('cardstate',days)] = df1['Amount']/df1['mean_amt_{}_{}'.format('cardstate',days)]
    
      df1['ratio_actual_max_amt_{}_{}'.format('card',days)] = df1['Amount']/df1['max_amt_{}_{}'.format('card',days)]
      df1['ratio_actual_max_amt_{}_{}'.format('merch',days)] = df1['Amount']/df1['max_amt_{}_{}'.format('merch',days)]
      df1['ratio_actual_max_amt_{}_{}'.format('cardmerch',days)] = df1['Amount']/df1['max_amt_{}_{}'.format('cardmerch',days)]
      df1['ratio_actual_max_amt_{}_{}'.format('cardzip',days)] = df1['Amount']/df1['max_amt_{}_{}'.format('cardzip',days)]
      df1['ratio_actual_max_amt_{}_{}'.format('cardstate',days)] = df1['Amount']/df1['max_amt_{}_{}'.format('cardstate',days)]
        
      df1['ratio_actual_median_amt_{}_{}'.format('card',days)] = df1['Amount']/df1['median_amt_{}_{}'.format('card',days)]
      df1['ratio_actual_median_amt_{}_{}'.format('merch',days)] = df1['Amount']/df1['median_amt_{}_{}'.format('merch',days)]
      df1['ratio_actual_median_amt_{}_{}'.format('cardmerch',days)] = df1['Amount']/df1['median_amt_{}_{}'.format('cardmerch',days)]
      df1['ratio_actual_median_amt_{}_{}'.format('cardzip',days)] = df1['Amount']/df1['median_amt_{}_{}'.format('cardzip',days)]
      df1['ratio_actual_median_amt_{}_{}'.format('cardstate',days)] = df1['Amount']/df1['median_amt_{}_{}'.format('cardstate',days)]
    
      df1['ratio_actual_sum_amt_{}_{}'.format('card',days)] = df1['Amount']/df1['sum_amt_{}_{}'.format('card',days)]
      df1['ratio_actual_sum_amt_{}_{}'.format('merch',days)] = df1['Amount']/df1['sum_amt_{}_{}'.format('merch',days)]
      df1['ratio_actual_sum_amt_{}_{}'.format('cardmerch',days)] = df1['Amount']/df1['sum_amt_{}_{}'.format('cardmerch',days)]
      df1['ratio_actual_sum_amt_{}_{}'.format('cardzip',days)] = df1['Amount']/df1['sum_amt_{}_{}'.format('cardzip',days)]
      df1['ratio_actual_sum_amt_{}_{}'.format('cardstate',days)] = df1['Amount']/df1['sum_amt_{}_{}'.format('cardstate',days)]


df1.to_csv("Amount_Variables_Final.csv")

pd.set_option('display.max_columns', 500)
df1.head()

df1 = pd.read_csv('Cleaned Data.csv')
df1['Date'] = pd.to_datetime(df1['Date'])
df1['Cardnum'] = df1['Cardnum'].astype(str)
df1['Merch zip'] = df1['Merch zip'].astype(str)
df1 = df1.set_index('Date')
df1.head()

#Calculates all the frequence variables - 25

df1 = df1.reset_index()
df1 = df1.sort_values(by = ['Cardnum', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
     df1['vel_{}_{}'.format('card',days)] = getattr(df1_index.groupby('Cardnum')['Recnum'].rolling(days),'count')().values
df1.sort_values(by = ['Recnum'],inplace = True)

df1 = df1.reset_index(drop = True)
df1 = df1.sort_values(by = ['Merchnum', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
     df1['vel_{}_{}'.format('merch',days)] = getattr(df1_index.groupby('Merchnum')['Recnum'].rolling(days),'count')().values    
df1.sort_values(by = ['Recnum'],inplace = True) 

df1 = df1.reset_index(drop = True)
df1 = df1.sort_values(by = ['Cardnum','Merchnum', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
     df1['vel_{}_{}'.format('cardmerch',days)] = getattr(df1_index.groupby(['Cardnum','Merchnum'])['Recnum'].rolling(days),'count')().values    
df1.sort_values(by = ['Recnum'],inplace = True)

df1 = df1.reset_index(drop = True)
df1 = df1.sort_values(by = ['Cardnum','Merch zip', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
     df1['vel_{}_{}'.format('cardzip',days)] = getattr(df1_index.groupby(['Cardnum','Merch zip'])['Recnum'].rolling(days),'count')().values    
df1.sort_values(by = ['Recnum'],inplace = True)  

df1 = df1.reset_index(drop = True)
df1 = df1.sort_values(by = ['Cardnum','Merch state', 'Date'])
df1_index = df1.set_index('Date')
for days in ['1d', '3d', '7d','14d','30d']:
     df1['vel_{}_{}'.format('cardstate',days)] = getattr(df1_index.groupby(['Cardnum','Merch state'])['Recnum'].rolling(days),'count')().values    
df1.sort_values(by = ['Recnum'],inplace = True)  


df1.to_csv("Frequency_Variables_Final.csv")
