==> Part3_Variable Creation.ipynb <==
# Importing necessary modules
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import datetime as dt
import warnings
warnings.filterwarnings("ignore")

amount = pd.read_csv("Amount_Variables_Final.csv")
frequency= pd.read_csv("Frequency_Variables_Final.csv")
frequency = frequency.iloc[:,10:36]

df1 = pd.concat([amount, frequency], axis=1, join='inner')
df1.head()

df1.shape

df1['Date'] = pd.to_datetime(df1['Date'])
df1['Cardnum'] = df1['Cardnum'].astype(str)
df1['Merch zip'] = df1['Merch zip'].astype(str)

import re
df1 = df1.rename(columns=lambda x: re.sub('1d','1',x))
df1 = df1.rename(columns=lambda x: re.sub('3d','3',x))
df1 = df1.rename(columns=lambda x: re.sub('7d','7',x))
df1 = df1.rename(columns=lambda x: re.sub('14d','14',x))
df1 = df1.rename(columns=lambda x: re.sub('30d','30',x))
df1.head()

# Calculates all the velocity change variables - 48 

for day in [7,14,30]:
    df1['rel_vel_vel_card_card_{}'.format(day)] = df1['vel_card_1']/ df1['vel_card_{}'.format(day)]
    df1['rel_vel_vel_card_merch_{}'.format(day)] = df1['vel_card_1']/ df1['vel_merch_{}'.format(day)]
    df1['rel_vel_amt_card_card_{}'.format(day)] = df1['vel_card_1']/ df1['sum_amt_card_{}'.format(day)]
    df1['rel_vel_amt_card_merch_{}'.format(day)] = df1['vel_card_1']/ df1['sum_amt_merch_{}'.format(day)]
    
    df1['rel_amt_vel_merch_card_{}'.format(day)] = df1['sum_amt_merch_1']/ df1['vel_card_{}'.format(day)]
    df1['rel_amt_vel_merch_merch_{}'.format(day)] = df1['sum_amt_merch_1']/ df1['vel_merch_{}'.format(day)]
    df1['rel_amt_amt_merch_card_{}'.format(day)] = df1['sum_amt_merch_1']/ df1['sum_amt_card_{}'.format(day)]
    df1['rel_amt_amt_merch_merch_{}'.format(day)] = df1['sum_amt_merch_1']/ df1['sum_amt_merch_{}'.format(day)]
    
    df1['rel_vel_vel_merch_card_{}'.format(day)] = df1['vel_merch_1']/ df1['vel_card_{}'.format(day)]
    df1['rel_vel_vel_merch_merch_{}'.format(day)] = df1['vel_merch_1']/ df1['vel_merch_{}'.format(day)]
    df1['rel_vel_amt_merch_card_{}'.format(day)] = df1['vel_merch_1']/ df1['sum_amt_card_{}'.format(day)]
    df1['rel_vel_amt_merch_merch_{}'.format(day)] = df1['vel_merch_1']/ df1['sum_amt_merch_{}'.format(day)]
    
    df1['rel_amt_vel_card_card_{}'.format(day)] = df1['sum_amt_card_1']/ df1['vel_card_{}'.format(day)]
    df1['rel_amt_vel_card_merch_{}'.format(day)] = df1['sum_amt_card_1']/ df1['vel_merch_{}'.format(day)]
    df1['rel_amt_amt_card_card_{}'.format(day)] = df1['sum_amt_card_1']/ df1['sum_amt_card_{}'.format(day)]
    df1['rel_amt_amt_card_merch_{}'.format(day)] = df1['sum_amt_card_1']/ df1['sum_amt_merch_{}'.format(day)]

df1.shape

df1.head()

# df1.to_csv("Velocity_Change_Variables_Final.csv")

# df1 = pd.read_csv('Cleaned Data.csv')
# df1['Date'] = pd.to_datetime(df1['Date'])
# df1['Cardnum'] = df1['Cardnum'].astype(str)
# df1['Merch zip'] = df1['Merch zip'].astype(str)
# df1.head()

# # Calculates all the days since variables - 5
# base_date = pd.to_datetime('2010-01-01')

# variable_name='dayssince_card'
# df1[variable_name] = df1.groupby('Cardnum')['Date'].diff().dt.days
# df1['temp'] = (df1.Date - base_date)
# df1['temp'] = df1.apply(lambda x: x.temp.days, axis=1)
# df1['temp'] = df1['temp'].astype(float)
# df1['temp'][df1['temp'] <20] = 19
# df1[variable_name] = df1[variable_name].replace(np.nan,df1['temp']+1)
# df1 = df1.drop('temp',axis = 1)

# variable_name='dayssince_merch'
# df1[variable_name] = df1.groupby('Merchnum')['Date'].diff().dt.days
# df1['temp'] = (df1.Date - base_date)
# df1['temp'] = df1.apply(lambda x: x.temp.days, axis=1)
# df1['temp'] = df1['temp'].astype(float)
# df1['temp'][df1['temp'] <20] = 19
# df1[variable_name] = df1[variable_name].replace(np.nan,df1['temp']+1)
# df1 = df1.drop('temp',axis = 1)

# variable_name='dayssince_cardmerch'
# df1[variable_name] = df1.groupby(['Cardnum','Merchnum'])['Date'].diff().dt.days
# df1['temp'] = (df1.Date - base_date)
# df1['temp'] = df1.apply(lambda x: x.temp.days, axis=1)
# df1['temp'] = df1['temp'].astype(float)
# df1['temp'][df1['temp'] <20] = 19
# df1[variable_name] = df1[variable_name].replace(np.nan,df1['temp']+1)
# df1 = df1.drop('temp',axis = 1)

# variable_name='dayssince_cardzip'
# df1[variable_name] = df1.groupby(['Cardnum','Merch zip'])['Date'].diff().dt.days
# df1['temp'] = (df1.Date - base_date)
# df1['temp'] = df1.apply(lambda x: x.temp.days, axis=1)
# df1['temp'] = df1['temp'].astype(float)
# df1['temp'][df1['temp'] <20] = 19
# df1[variable_name] = df1[variable_name].replace(np.nan,df1['temp']+1)
# df1 = df1.drop('temp',axis = 1)

# variable_name='dayssince_cardstate'
# df1[variable_name] = df1.groupby(['Cardnum','Merch state'])['Date'].diff().dt.days
# df1['temp'] = (df1.Date - base_date)
# df1['temp'] = df1.apply(lambda x: x.temp.days, axis=1)
# df1['temp'] = df1['temp'].astype(float)
# df1['temp'][df1['temp'] <20] = 19
# df1[variable_name] = df1[variable_name].replace(np.nan,df1['temp']+1)
# df1 = df1.drop('temp',axis = 1)

# # Subtracts days since variable from max to keep it consistent (higher number means bad and lower number means good now)
# days_since_columns = ['dayssince_card','dayssince_merch','dayssince_cardmerch','dayssince_cardzip','dayssince_cardstate']

# for i in days_since_columns: 
#     maximum = df1[i].max()
#     df1[i] = maximum - df1[i]

# df1.shape

# df1.head()

# df1.to_csv("Dayssince_Variables_Final.csv")
